requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
google-genai>=1.59.0
google-cloud-aiplatform>=1.133.0
pillow>=12.0.0
pydantic>=2.9.0
pydantic-settings>=2.3.0
python-multipart==0.0.6
aiofiles==23.2.1
python-dotenv==1.0.0
